refactor utils class
