+++ app/src/main/java/com/alexanderpodkopaev/androidacademyproject/utils/OffsetItemDecoration.kt
@@ -1,16 +1,18 @@
-package com.alexanderpodkopaev.androidacademyproject
+package com.alexanderpodkopaev.androidacademyproject.utils
 
 import android.graphics.Rect
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 
-class RightOffsetItemDecoration(val offset: Int) : RecyclerView.ItemDecoration() {
+class OffsetItemDecoration(val offset: Int) : RecyclerView.ItemDecoration() {
     override fun getItemOffsets(
         outRect: Rect,
         view: View,
         parent: RecyclerView,
         state: RecyclerView.State
     ) {
-        outRect.right = offset
+        outRect.left = offset / 2
+        outRect.right = offset / 2
+        outRect.bottom = offset
     }
 }
