+++ app/src/main/java/com/alexanderpodkopaev/androidacademyproject/utils/RightOffsetItemDecoration.kt
@@ -1,18 +1,16 @@
-package com.alexanderpodkopaev.androidacademyproject
+package com.alexanderpodkopaev.androidacademyproject.utils
 
 import android.graphics.Rect
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 
-class OffsetItemDecoration(val offset: Int) : RecyclerView.ItemDecoration() {
+class RightOffsetItemDecoration(val offset: Int) : RecyclerView.ItemDecoration() {
     override fun getItemOffsets(
         outRect: Rect,
         view: View,
         parent: RecyclerView,
         state: RecyclerView.State
     ) {
-        outRect.left = offset / 2
-        outRect.right = offset / 2
-        outRect.bottom = offset
+        outRect.right = offset
     }
 }
